Robot-behavior/scripts/project_status.py
#!/usr/bin/env python3
"""
🚀 Robot Behavior Simulator - Project Organization Summary

This script provides a comprehensive overview of the project organization
and package-ready status.
"""

import os
import sys
from pathlib import Path
import subprocess

def print_banner(message):
    """Print a formatted banner."""
    print("\n" + "="*60)
    print(f"🤖 {message}")
    print("="*60)

def check_file_exists(filepath, description):
    """Check if a file exists and print status."""
    if Path(filepath).exists():
        print(f"✅ {description}: {filepath}")
        return True
    else:
        print(f"❌ Missing {description}: {filepath}")
        return False

def show_project_structure():
    """Display the organized project structure."""
    print_banner("Project Structure Overview")
    
    root = Path(".")
    important_files = {
        # Core package files
        "pyproject.toml": "Modern Python packaging configuration",
        "setup.py": "Legacy setup script (backup)",
        "README.md": "Project documentation",
        "LICENSE": "MIT license file",
        "requirements.txt": "Dependencies specification",
        "MANIFEST.in": "Package data inclusion rules",
        
        # Package structure
        "robot_behavior/__init__.py": "Main package initialization",
        "robot_behavior/minimal_api.py": "Student-facing API",
        "robot_behavior/core/robot.py": "Core robot logic",
        "robot_behavior/simulator/enhanced_simulator.py": "Visual simulation engine",
        "robot_behavior/assets/dog.png": "Robot graphics asset",
        
        # Documentation
        "docs/MINIMAL_API_GUIDE.md": "Complete API documentation",
        
        # Examples and tests
        "examples/minimal_api_examples.py": "Working example code",
        "tests/test_robot_behavior.py": "Comprehensive test suite",
        
        # Build and scripts
        "scripts/build_package.py": "Build automation script",
        "scripts/restore_working.py": "State restoration utility",
    }
    
    all_exist = True
    for filepath, description in important_files.items():
        if not check_file_exists(filepath, description):
            all_exist = False
    
    if all_exist:
        print("\n🎉 All critical files present!")
    else:
        print("\n⚠️  Some files are missing - see above")
    
    return all_exist

def show_package_info():
    """Display package metadata information."""
    print_banner("Package Information")
    
    try:
        # Import package to get metadata
        sys.path.insert(0, str(Path.cwd()))
        import robot_behavior
        
        print(f"📦 Package Name: robot-behavior-simulator")
        print(f"🏷️  Version: {robot_behavior.__version__}")
        print(f"👥 Author: {robot_behavior.__author__}")
        print(f"📧 Email: {robot_behavior.__email__}")
        print(f"📄 License: {robot_behavior.__license__}")
        print(f"🔗 URL: {robot_behavior.__url__}")
        
        # Check available functions
        print(f"\n📚 Available Functions:")
        for name in robot_behavior.__all__:
            if hasattr(robot_behavior, name):
                obj = getattr(robot_behavior, name)
                obj_type = "function" if callable(obj) else "constant"
                print(f"  • {name} ({obj_type})")
        
        print("✅ Package imports successfully!")
        return True
        
    except Exception as e:
        print(f"❌ Package import failed: {e}")
        return False

def check_dependencies():
    """Check if all dependencies are available."""
    print_banner("Dependency Check")
    
    dependencies = {
        "Pillow": "Image processing (required)",
        "tkinter": "GUI framework (built-in)",
        "pytest": "Testing framework (optional)",
        "build": "Package building (optional)",
        "twine": "Package uploading (optional)",
    }
    
    all_deps_ok = True
    for dep_name, description in dependencies.items():
        try:
            if dep_name == "tkinter":
                import tkinter
            elif dep_name == "Pillow":
                from PIL import Image, ImageTk, ImageDraw
            elif dep_name == "pytest":
                import pytest
            elif dep_name == "build":
                import build
            elif dep_name == "twine":
                import twine
            
            print(f"✅ {dep_name}: {description}")
        except ImportError:
            if dep_name in ["pytest", "build", "twine"]:
                print(f"⚠️  {dep_name}: {description} (optional - not installed)")
            else:
                print(f"❌ {dep_name}: {description} (MISSING - required)")
                all_deps_ok = False
    
    return all_deps_ok

def show_build_commands():
    """Show available build and packaging commands."""
    print_banner("Build & Package Commands")
    
    print("🧹 Cleaning:")
    print("  python scripts/build_package.py clean")
    
    print("\n🧪 Testing:")
    print("  python scripts/build_package.py test")
    print("  python -m pytest tests/ -v")
    
    print("\n📦 Building:")
    print("  python scripts/build_package.py build")
    print("  python -m build")
    
    print("\n✅ Quality Check:")
    print("  python scripts/build_package.py check")
    print("  python -m twine check dist/*")
    
    print("\n🚀 Publishing:")
    print("  python scripts/build_package.py upload-test  # TestPyPI")
    print("  python scripts/build_package.py upload      # PyPI")
    
    print("\n⚡ Complete Pipeline:")
    print("  python scripts/build_package.py all")

def show_usage_examples():
    """Show usage examples for the package."""
    print_banner("Usage Examples")
    
    print("🟢 Basic Usage:")
    print("""
from robot_behavior import create_robot_program

program = create_robot_program(5, 5, 0, 0)
program.robot.move('right')
program.robot.move('up')
program.start()
""")
    
    print("🟡 With Animation:")
    print("""
from robot_behavior import create_robot_program, run_with_visualization

program = create_robot_program(8, 6, 2, 2)
def my_moves():
    program.robot.move('right')
    program.robot.move('up')
    program.robot.move('left')

run_with_visualization(program, my_moves, move_delay=0.8)
""")
    
    print("🔴 Maze Navigation:")
    print("""
from robot_behavior import create_robot_program, SIMPLE_MAZE

program = create_robot_program()
program.load_maze(SIMPLE_MAZE)

# Navigate through maze...
program.robot.move('right')
program.robot.move('up')
program.start()
""")

def main():
    """Main function to show project organization status."""
    print("🤖 Robot Behavior Simulator - Project Organization Report")
    print("="*70)
    
    # Check project structure
    structure_ok = show_project_structure()
    
    # Check package info
    package_ok = show_package_info()
    
    # Check dependencies
    deps_ok = check_dependencies()
    
    # Show build commands
    show_build_commands()
    
    # Show usage examples
    show_usage_examples()
    
    # Final status
    print_banner("Package Ready Status")
    
    if structure_ok and package_ok and deps_ok:
        print("🎉 PROJECT IS PACKAGE-READY! 🎉")
        print()
        print("✅ All files organized correctly")
        print("✅ Package imports successfully")
        print("✅ Dependencies satisfied")
        print("✅ Build scripts available")
        print("✅ Comprehensive documentation")
        print("✅ Test suite included")
        print()
        print("🚀 Ready for:")
        print("  • Local development and testing")
        print("  • Package building and distribution")
        print("  • PyPI publishing")
        print("  • Educational deployment")
        print()
        print("🎯 Next steps:")
        print("  1. Run: python scripts/build_package.py all")
        print("  2. Test: pip install dist/*.whl")
        print("  3. Publish: python scripts/build_package.py upload-test")
        
    else:
        print("⚠️  PROJECT NEEDS ATTENTION")
        print()
        if not structure_ok:
            print("❌ Fix missing files")
        if not package_ok:
            print("❌ Fix package import issues")
        if not deps_ok:
            print("❌ Install missing dependencies")
        print()
        print("🔧 Run this script again after fixes")

if __name__ == "__main__":
    main()
